requirements.txt
tensorflow>=2.8.0
numpy>=1.19.5
pandas>=1.3.0
pillow>=8.3.0
matplotlib>=3.4.0
tqdm>=4.61.0
scikit-learn~=1.6.0
pytest~=8.3.4